test(daemon): split runtime ingress fixtures and align fast contract fixtures
