+++ packages/daemon/test/runtime-spawn-ingress.integration.test.ts
@@ -1,42 +1,30 @@
 // harness-test-tier: integration
 import assert from "node:assert/strict";
-import { execFileSync, spawn } from "node:child_process";
+import { spawn } from "node:child_process";
 import { mkdirSync, mkdtempSync, readFileSync, realpathSync, rmSync, writeFileSync } from "node:fs";
-import { hostname, tmpdir } from "node:os";
+import { tmpdir } from "node:os";
 import path from "node:path";
 import test from "node:test";
-import { makeTaskEventStore, makeTaskProjection, type AgentDefinitionSnapshot } from "../../kernel/src/index.ts";
-import { type RuntimeInstallationWitness } from "../src/agent-runtime-instances.ts";
+import { makeTaskEventStore, makeTaskProjection } from "../../kernel/src/index.ts";
 import { localUserDaemonEndpoint } from "../src/client/local-daemon-target.ts";
 import { openDaemonHost } from "../src/daemon-host.ts";
 import { createJsonRpcProtocolServer } from "../src/protocol/json-rpc-server.ts";
-import { currentDaemonProtocolVersion } from "../src/protocol/version.ts";
 import { createUnixSocketTransportServer } from "../src/transport/unix-socket.ts";
 import { writeProviderExecutable } from "./fixtures/runtime-stub.ts";
 import { registerBootstrappedDaemonRepo as registerDaemonRepo } from "./repo-settings.fixture.ts";
 import { createRealizedTaskPlanFixture } from "../../../tools/fixtures/task-plan.mjs";
-
-const cli = path.resolve("packages/cli/src/index.ts");
-const definition: AgentDefinitionSnapshot = {
-  schema: "agent-definition-snapshot/v1",
-  configVersion: 1,
-  instanceId: "codex-review",
-  installationId: "installation-codex",
-  kindId: "codex",
-  providerId: "openai",
-  model: "gpt-5.6-sol",
-  reasoningEffort: "high",
-  fast: false,
-  baseUrl: "https://api.example.test/",
-  authMode: "api-key",
-};
-const installation: RuntimeInstallationWitness = {
-  installationId: definition.installationId,
-  kindId: definition.kindId,
-  executablePath: "/opt/witnessed/codex",
-  version: "1.0.0",
-  observedAt: "2026-08-14T00:00:00.000Z",
-};
+import {
+  definition,
+  installation,
+  initIngressRepo,
+  rpc,
+  rpcAttach,
+  eventually,
+  eventuallyValue,
+  writeProviderStub,
+  installationFixture,
+  spawnCli,
+} from "./fixtures/runtime-ingress.ts";
 
 test("daemon ingress preserves executor-scoped task-bound runtime spawn", async (t) => {
   const parent = mkdtempSync(path.join(tmpdir(), "ha-runtime-spawn-ingress-")),
@@ -1392,191 +1380,6 @@ test("daemon ingress cancellation is explicit and idempotent for an active runti
   }
 });
 
-test("daemon ingress passes Claude effort and Codex fast through to the witnessed CLIs", async () => {
-  const parent = mkdtempSync(path.join(tmpdir(), "ha-runtime-claude-effort-ingress-")),
-    root = path.join(parent, "repo"),
-    userRoot = path.join(parent, "user"),
-    repoId = "runtime-claude-effort-ingress",
-    uid = process.getuid?.() ?? 0,
-    argsPath = path.join(parent, "claude-args.json"),
-    codexArgsPath = path.join(parent, "codex-args.json"),
-    executablePath = writeProviderStub(path.join(parent, "claude-stub"), "claude", argsPath),
-    codexExecutablePath = writeProviderStub(path.join(parent, "codex-stub"), "codex", codexArgsPath),
-    claudeInstallation = {
-      installationId: "installation-claude-effort",
-      kindId: "claude" as const,
-      executablePath,
-      version: "2.1.251",
-      observedAt: "2026-08-30T00:00:00.000Z",
-    },
-    codexInstallation = {
-      installationId: "installation-codex-fast",
-      kindId: "codex" as const,
-      executablePath: codexExecutablePath,
-      version: "0.150.1",
-      observedAt: "2026-08-30T00:00:00.000Z",
-    },
-    auth = {
-      transportKind: "unix-socket",
-      unixSocketOwnerBoundary: {
-        ownerUid: uid,
-        source: "unix-socket-filesystem-owner-boundary",
-      },
-    } as const;
-  initIngressRepo(root, uid);
-  registerDaemonRepo({ canonicalRoot: root, repoId, userRoot, createConvenienceLinks: false });
-  const host = await openDaemonHost({
-      daemonId: "runtime-claude-effort-ingress",
-      userRoot,
-      runtimeDiscover: () => [claudeInstallation, codexInstallation],
-    }),
-    endpoint = localUserDaemonEndpoint(userRoot, "runtime-claude-effort-ingress"),
-    transport = createUnixSocketTransportServer({
-      daemonId: "runtime-claude-effort-ingress",
-      socketPath: endpoint,
-      createProtocolServer: (authContext, emit) =>
-        createJsonRpcProtocolServer({ host, build: { commit: null }, authContext, emit }),
-    });
-  await transport.start();
-  try {
-    await host.attachmentsSettled();
-    host.runtimeInstance(
-      "daemon.runtimeInstance.create",
-      {
-        instanceId: "claude-effort-provider",
-        name: "Claude effort provider",
-        kindId: "claude",
-        installationId: claudeInstallation.installationId,
-        providerId: "anthropic",
-        models: ["claude-fable-5"],
-        claude: {},
-        authMode: "subscription",
-      },
-      auth,
-    );
-    host.runtimeInstance(
-      "daemon.runtimeInstance.create",
-      {
-        instanceId: "codex-fast-provider",
-        name: "Codex fast provider",
-        kindId: "codex",
-        installationId: codexInstallation.installationId,
-        providerId: "openai",
-        models: ["gpt-5.6-sol"],
-        codex: {},
-        authMode: "subscription",
-      },
-      auth,
-    );
-    const childEnv = {
-        ...process.env,
-        HARNESS_DAEMON_USER_ROOT: userRoot,
-        HARNESS_DAEMON_ID: "runtime-claude-effort-ingress",
-        HARNESS_DAEMON_REPO_ID: repoId,
-        HARNESS_DAEMON_ENDPOINT: endpoint,
-      },
-      expectedArgs = [
-        "-p",
-        "--verbose",
-        "--output-format",
-        "stream-json",
-        "--permission-mode",
-        "bypassPermissions",
-        "--model",
-        "claude-fable-5",
-        "--effort",
-        "high",
-      ],
-      spawned = await spawnCli(
-        [
-          "--root",
-          root,
-          "--json",
-          "runtime",
-          "run",
-          "claude-effort-provider",
-          "--effort",
-          "high",
-          "--prompt",
-          "probe",
-          "--no-stream",
-        ],
-        childEnv,
-      );
-    assert.equal(spawned.status, 0, `${spawned.stderr}\n${spawned.stdout}`);
-    assert.deepEqual(JSON.parse(readFileSync(argsPath, "utf8")), expectedArgs);
-    const fastSpawned = await spawnCli(
-      ["--root", root, "--json", "runtime", "run", "codex-fast-provider", "--fast", "--prompt", "probe", "--no-stream"],
-      childEnv,
-    );
-    assert.equal(fastSpawned.status, 0, `${fastSpawned.stderr}\n${fastSpawned.stdout}`);
-    assert.deepEqual(JSON.parse(readFileSync(codexArgsPath, "utf8")), [
-      "exec",
-      "--json",
-      "--sandbox",
-      "danger-full-access",
-      "--model",
-      "gpt-5.6-sol",
-      "--config",
-      'service_tier="fast"',
-      "-",
-    ]);
-    const unsupported = await spawnCli(
-      [
-        "--root",
-        root,
-        "--json",
-        "runtime",
-        "run",
-        "claude-effort-provider",
-        "--fast",
-        "--prompt",
-        "reject",
-        "--no-stream",
-      ],
-      childEnv,
-    );
-    assert.notEqual(unsupported.status, 0);
-    assert.equal((JSON.parse(unsupported.stdout) as Record<string, unknown>).code, "invalid_runtime_fast");
-    const rejected = await spawnCli(
-      [
-        "--root",
-        root,
-        "--json",
-        "runtime",
-        "run",
-        "claude-effort-provider",
-        "--effort",
-        "turbo",
-        "--prompt",
-        "reject",
-        "--no-stream",
-      ],
-      childEnv,
-    );
-    assert.equal(rejected.status, 2, `${rejected.stderr}\n${rejected.stdout}`);
-    assert.equal((JSON.parse(rejected.stdout) as Record<string, unknown>).code, "invalid_runtime_effort");
-    const rejectedIngress = await rpc(host, auth, "repo.agentRuntime.spawn", {
-      repo: { repoId },
-      payload: {
-        runtimeInstanceId: "claude-effort-provider",
-        cwd: { scope: "repo-root" },
-        prompt: "reject",
-        effort: "turbo",
-        taskId: null,
-        idempotencyKey: "claude-effort-turbo",
-      },
-    });
-    assert.equal(rejectedIngress.outcome, "op_rejected", JSON.stringify(rejectedIngress));
-    assert.equal(rejectedIngress.code, "invalid_runtime_effort");
-    assert.deepEqual(JSON.parse(readFileSync(argsPath, "utf8")), expectedArgs);
-  } finally {
-    await transport.stop();
-    await host.close();
-    rmSync(parent, { recursive: true, force: true });
-  }
-});
-
 test("agy consumes only its closed stream-json event protocol", async () => {
   const parent = mkdtempSync(path.join(tmpdir(), "ha-runtime-agy-events-")),
     root = path.join(parent, "repo"),
@@ -1723,231 +1526,3 @@ test("agy consumes only its closed stream-json event protocol", async () => {
     rmSync(parent, { recursive: true, force: true });
   }
 });
-
-function initIngressRepo(root: string, uid: number): void {
-  mkdirSync(path.join(root, "harness"), { recursive: true });
-  git(root, "init", "-q");
-  git(root, "config", "user.name", "Spawn Test");
-  git(root, "config", "user.email", "spawn@example.invalid");
-  writeFileSync(
-    path.join(root, "harness/harness.yaml"),
-    "schema: harness-anything/v1\nname: runtime-spawn-ingress\nlayout:\n  authoredRoot: harness\n  localRoot: .harness\n",
-  );
-  writeFileSync(
-    path.join(root, "harness/people.yaml"),
-    `${JSON.stringify({ schema: "harness-people/v1", people: [{ personId: "owner", displayName: "Owner", roles: ["owner"], credentials: [{ kind: "unix-socket-owner-boundary", issuer: `host:${hostname()}`, subject: String(uid) }] }], roles: [{ roleId: "owner", commandClasses: ["repo-read", "repo-write"] }] }, null, 2)}\n`,
-  );
-  git(root, "add", "harness");
-  git(root, "commit", "-qm", "fixture");
-}
-async function rpc(
-  host: Awaited<ReturnType<typeof openDaemonHost>>,
-  auth: Parameters<Awaited<ReturnType<typeof openDaemonHost>>["run"]>[2],
-  method: string,
-  params: Record<string, unknown>,
-): Promise<Record<string, unknown>> {
-  const server = createJsonRpcProtocolServer({
-    host,
-    build: { commit: null },
-    authContext: auth,
-    emit: async () => undefined,
-  });
-  try {
-    await server.handle({
-      jsonrpc: "2.0",
-      id: 1,
-      method: "protocol.hello",
-      params: { protocolVersion: currentDaemonProtocolVersion },
-    });
-    const response = await server.handle({
-      jsonrpc: "2.0",
-      id: 2,
-      method,
-      params,
-    });
-    assert.ok(response && !Array.isArray(response) && "result" in response);
-    return (response as { result: Record<string, unknown> }).result;
-  } finally {
-    server.close();
-  }
-}
-async function rpcAttach(
-  host: Awaited<ReturnType<typeof openDaemonHost>>,
-  auth: Parameters<Awaited<ReturnType<typeof openDaemonHost>>["run"]>[2],
-  repoId: string,
-  runtimeSessionId: string,
-  frames: Record<string, unknown>[],
-): Promise<{ readonly close: () => void }> {
-  const server = createJsonRpcProtocolServer({
-    host,
-    build: { commit: null },
-    authContext: auth,
-    emit: async (_method, params) => {
-      frames.push(params);
-    },
-  });
-  await server.handle({
-    jsonrpc: "2.0",
-    id: 1,
-    method: "protocol.hello",
-    params: { protocolVersion: currentDaemonProtocolVersion },
-  });
-  const response = await server.handle({
-    jsonrpc: "2.0",
-    id: 2,
-    method: "repo.agentRuntime.attach",
-    params: {
-      repo: { repoId },
-      payload: { runtimeSessionId, afterCursor: "stream:0" },
-    },
-  });
-  assert.ok(response && !Array.isArray(response) && "result" in response);
-  assert.equal((response as { result: { ok: boolean } }).result.ok, true, JSON.stringify(response));
-  return { close: server.close };
-}
-async function eventually(check: () => boolean | Promise<boolean>): Promise<void> {
-  await eventuallyValue(async () => ((await check()) ? true : null));
-}
-async function eventuallyValue<T>(read: () => T | null | Promise<T | null>): Promise<T> {
-  for (let attempt = 0; attempt < 500; attempt += 1) {
-    const value = await read();
-    if (value !== null) return value;
-    await new Promise((resolve) => setTimeout(resolve, 10));
-  }
-  throw new Error("runtime provider event did not arrive");
-}
-function writeProviderStub(target: string, kindId: "claude" | "codex", argsTarget?: string): string {
-  const lines =
-    kindId === "claude"
-      ? [
-          {
-            type: "system",
-            subtype: "init",
-            session_id: "claude-provider-session",
-          },
-          {
-            type: "assistant",
-            session_id: "claude-provider-session",
-            message: {
-              content: [
-                { type: "text", text: "claude live content" },
-                {
-                  type: "tool_use",
-                  id: "write-1",
-                  name: "Write",
-                  input: { file_path: "result.txt", content: "written" },
-                },
-              ],
-            },
-          },
-          {
-            type: "user",
-            session_id: "claude-provider-session",
-            tool_use_result: { type: "create", filePath: "result.txt" },
-            message: {
-              content: [
-                {
-                  type: "tool_result",
-                  tool_use_id: "write-1",
-                  content: "created",
-                },
-              ],
-            },
-          },
-          {
-            type: "result",
-            subtype: "success",
-            is_error: false,
-            session_id: "claude-provider-session",
-            result: "claude final result",
-            permission_denials: [],
-          },
-        ]
-      : [
-          { type: "thread.started", thread_id: "codex-provider-session" },
-          {
-            type: "item.completed",
-            item: {
-              id: "item-1",
-              type: "agent_message",
-              text: "codex live content",
-              credentialRef: "credential-secret",
-              executablePath: "/provider/private",
-              apiToken: "sk-provider-secret",
-            },
-          },
-          {
-            type: "item.completed",
-            item: {
-              id: "write-1",
-              type: "file_change",
-              changes: [{ path: "result.txt", kind: "add" }],
-              status: "completed",
-            },
-          },
-          {
-            type: "item.completed",
-            item: {
-              id: "item-2",
-              type: "agent_message",
-              text: "codex final result",
-            },
-          },
-          {
-            type: "turn.completed",
-            usage: { input_tokens: 1, output_tokens: 1 },
-          },
-        ];
-  const structuredFlag =
-    kindId === "claude"
-      ? `process.argv.includes("--output-format") && process.argv.includes("stream-json") && process.argv.includes("--verbose")`
-      : `process.argv[2] === "exec" && process.argv.includes("--json")`;
-  const recordArgs = argsTarget
-    ? `fs.writeFileSync(${JSON.stringify(argsTarget)}, JSON.stringify(process.argv.slice(2)));\n`
-    : "";
-  return writeProviderExecutable(
-    target,
-    `import fs from "node:fs";\nconst auth = process.argv[2] === "auth" || process.argv[2] === "login";\nif (auth) process.exit(0);\n${recordArgs}if (!(${structuredFlag})) process.exit(9);\nconst prompt = fs.readFileSync(0, "utf8"), secret = "sk-runtime-secret-1234567890";\nif (prompt === "failure:empty") process.exit(1);\nelse if (prompt === "failure:secret") process.stderr.write("OPENAI_API_KEY=" + secret + "\\n", () => process.exit(1));\nelse if (prompt === "failure:structured") process.stdout.write([JSON.stringify({ type: "thread.started", thread_id: "codex-provider-session" }), JSON.stringify({ type: "turn.failed", error: { message: "structured provider failure", apiToken: secret } })].join("\\n") + "\\n", () => process.exit(1));\nelse if (prompt === "permission-denied") process.stdout.write([JSON.stringify({ type: "system", subtype: "init", session_id: "claude-provider-session" }), JSON.stringify({ type: "assistant", session_id: "claude-provider-session", message: { content: [{ type: "tool_use", id: "denied-write", name: "Write", input: { file_path: "/tmp/outside", content: "denied" } }] } }), JSON.stringify({ type: "result", subtype: "success", is_error: false, session_id: "claude-provider-session", result: "write denied", permission_denials: [{ tool_name: "Write", tool_use_id: "denied-write" }] })].join("\\n") + "\\n");\nelse { let emitted = ${JSON.stringify(lines)}; if (prompt === "read-only") emitted = [{ type: "thread.started", thread_id: "codex-provider-session" }, { type: "item.completed", item: { id: "inspect", type: "command_execution", command: "git status --short", aggregated_output: "", exit_code: 0, status: "completed" } }, { type: "item.completed", item: { id: "message", type: "agent_message", text: "read-only final result" } }, { type: "turn.completed", usage: { input_tokens: 1, output_tokens: 1 } }]; else if (prompt === "no-action") emitted = [{ type: "thread.started", thread_id: "codex-provider-session" }, { type: "item.completed", item: { id: "message", type: "agent_message", text: "no-action final result" } }, { type: "turn.completed", usage: { input_tokens: 1, output_tokens: 1 } }]; else if (prompt === "no-write") emitted.splice(1, 0, { type: "item.completed", item: { id: "inspect", type: "command_execution", command: "git status --short", aggregated_output: "", exit_code: 0, status: "completed" } }, { type: "item.updated", item: { id: "plan", type: "todo_list", items: [{ text: "locate cause", status: "completed" }, { text: "write fix", status: "in_progress" }] } }); emitted.forEach((line, index) => setTimeout(() => console.log(JSON.stringify(line)), index * 40)); }\n`,
-  );
-}
-function installationFixture(kindId: "claude" | "codex", executablePath: string): RuntimeInstallationWitness {
-  return {
-    installationId: `installation-${kindId}`,
-    kindId,
-    executablePath,
-    version: "1.0.0",
-    observedAt: "2026-08-19T00:00:00.000Z",
-  };
-}
-function spawnCli(
-  args: readonly string[],
-  env: NodeJS.ProcessEnv,
-): Promise<{
-  readonly status: number | null;
-  readonly stdout: string;
-  readonly stderr: string;
-}> {
-  return new Promise((resolve, reject) => {
-    const child = spawn(process.execPath, [cli, ...args], {
-      env,
-      stdio: ["ignore", "pipe", "pipe"],
-    });
-    let stdout = "",
-      stderr = "";
-    const timeout = setTimeout(() => child.kill(), 10_000);
-    child.stdout.setEncoding("utf8").on("data", (chunk) => {
-      stdout += chunk;
-    });
-    child.stderr.setEncoding("utf8").on("data", (chunk) => {
-      stderr += chunk;
-    });
-    child.once("error", reject);
-    child.once("close", (status) => {
-      clearTimeout(timeout);
-      resolve({ status, stdout, stderr });
-    });
-  });
-}
-function git(root: string, ...args: string[]): void {
-  execFileSync("git", ["-C", root, ...args]);
-}
