+++ packages/daemon/test/runtime-spawn-lifecycle.integration.test.ts
@@ -158,7 +158,7 @@ test("runtime spawn publishes a canonical session and makes it visible in overvi
           error.code === "invalid_runtime_spawn" &&
           error.message ===
             'Runtime spawn payload contains an unknown field "permission_mode"; allowed fields: "runtimeInstanceId", ' +
-              '"dispatchId", "agentId", "targetAgentId", "squadId", "model", "effort", "permissionMode", "cwd", ' +
+              '"dispatchId", "agentId", "targetAgentId", "squadId", "model", "effort", "fast", "permissionMode", "cwd", ' +
               '"prompt", "promptSource", "missionName", "onExitCommand", "taskId", "idempotencyKey", ' +
               '"providerSessionId".',
       );
